refactor (Server template): Use 'ServerBusinessLogic' name for service classes processing webhooks business logic
