+++ templates/server/src/main/java/com/mycompany/app/numbers/Controller.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class Controller {
 
   private final SinchClient sinchClient;
-  private final WebhooksBusinessLogic webhooksBusinessLogic;
+  private final ServerBusinessLogic webhooksBusinessLogic;
 
   @Autowired
-  public Controller(SinchClient sinchClient, WebhooksBusinessLogic webhooksBusinessLogic) {
+  public Controller(SinchClient sinchClient, ServerBusinessLogic webhooksBusinessLogic) {
     this.sinchClient = sinchClient;
     this.webhooksBusinessLogic = webhooksBusinessLogic;
   }
