+++ templates/server/src/main/java/com/mycompany/app/numbers/ServerBusinessLogic.java
@@ -4,10 +4,10 @@ import com.sinch.sdk.domains.numbers.models.v1.webhooks.NumberEvent;
 import java.util.logging.Logger;
 import org.springframework.stereotype.Component;
 
-@Component("NumbersWebhooksBusinessLogic")
-public class WebhooksBusinessLogic {
+@Component("NumbersServerBusinessLogic")
+public class ServerBusinessLogic {
 
-  private static final Logger LOGGER = Logger.getLogger(WebhooksBusinessLogic.class.getName());
+  private static final Logger LOGGER = Logger.getLogger(ServerBusinessLogic.class.getName());
 
   public void numbersEvent(NumberEvent event) {
